% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doAnalysis}
\alias{doAnalysis}
\title{Analyse SMART data}
\usage{
doAnalysis(
  data,
  scheme,
  u = c(0, 0, 0, 0, 1, 1, 1, 1, 1, 1),
  myopic = F,
  tuning = 0
)
}
\arguments{
\item{data}{A dataframe containing simulated SMART data.}

\item{scheme}{A dataframe that defines the structure of the SMART in terms of stages (`stage`),
treatments and interactions (`a1`, `a2`, `a1a2`), the intermediate outcome (`y1`), and initial randomisation
probabilities (`rho`), and event rates for the specified treatments (`pi`).}

\item{u}{A vector of utilities.}

\item{myopic}{A logical. Does this analysis use 'myopic' response adaptive randomisation (as opposed to 'dynamic' response adaptive randomisation)? Defaults to FALSE.}

\item{tuning}{A real values scalar used to define whether any randomisation is equal (`tuning` equals zero) or adaptive (`tuning` greater than zero, with adaptive algorithm being 'greedier' for higher values). Defaults to zero.}
}
\value{

}
\description{
\code{doAnalysis} analyses simulated SMART data using a Q-learning decision-theoretic model implemented in Rstan. Requires the Rstan model (`qModel`) to be compiled before running.
}
\examples{

For a two-stage, two-treatment, degenerate (i.e. participants exit the trial if event) we simulate a participant
with equal randomisation and event rates as follows:

scheme <- data.frame(
stage = c(   1,    1,   2,   2,   2,   2),
int   = c(   1,    1,   1,   1,   1,   1),
a1    = c(   0,    1,   0,   0,   1,   1),
y1    = c(   0,    0,   1,   1,   1,   1),
a2    = c(  NA,   NA,   0,   1,   0,   1),
a1a2  = c(  NA,   NA,   0,   0,   0,   1),
rho   = c( 0.5,  0.5, 0.5, 0.5, 0.5, 0.5),
pi    = c( 0.8,  0.7, 0.5, 0.4, 0.5, 0.4))

data <- lapply(1:1000, function(x) simulatePatients(scheme))
data <- do.call(rbind, data)

doAnalysis(data, scheme)

}
