% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getKnownUtility}
\alias{getKnownUtility}
\title{Get known utilities.}
\usage{
getKnownUtility(scheme, u = c(0, 0, 0, 0, 1, 1, 1, 1, 1, 1))
}
\arguments{
\item{scheme}{A dataframe that defines the structure of the SMART in terms of stages (`stage`),
treatments and interactions (`a1`, `a2`, `a1a2`), the intermediate outcome (`y1`), and initial randomisation
probabilities (`rho`), and event rates for the specified treatments (`pi`).}

\item{u}{A vector of utilities.}
}
\value{

}
\description{
\code{getKnownUtilities} obtains the true utilities for a given utility set and trial scheme.
}
\examples{

For a two-stage, two-treatment, degenerate (i.e. participants exit the trial if event) we simulate a participant
with equal randomisation and event rates as follows:

scheme <- data.frame(
stage = c(   1,    1,   2,   2,   2,   2),
int   = c(   1,    1,   1,   1,   1,   1),
a1    = c(   0,    1,   0,   0,   1,   1),
y1    = c(   0,    0,   1,   1,   1,   1),
a2    = c(  NA,   NA,   0,   1,   0,   1),
a1a2  = c(  NA,   NA,   0,   0,   0,   1),
rho   = c( 0.5,  0.5, 0.5, 0.5, 0.5, 0.5),
pi    = c( 0.8,  0.7, 0.5, 0.4, 0.5, 0.4))

getKnownUtility(scheme)

}
