% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulateTrial}
\alias{simulateTrial}
\title{Simulate a decision-theoretic SMART with response adaptive randomisation.}
\usage{
simulateTrial(scheme, interimSize, interimNumber, tuning = 0, myopic = F, u)
}
\arguments{
\item{scheme}{A dataframe that defines the structure of the SMART in terms of stages (`stage`),
treatments and interactions (`a1`, `a2`, `a1a2`), the intermediate outcome (`y1`), and initial randomisation
probabilities (`rho`), and event rates for the specified treatments (`pi`).}

\item{interimSize}{A positive integer indicating the size of (equal sized and complementary) analysis cohorts used in the analysis.}

\item{interimNumber}{A positive integer indicating the number of interim analyses to be performed (include the final analysis).}

\item{tuning}{A real values scalar used to define whether any randomisation is equal (`tuning` equals zero) or adaptive (`tuning` greater than zero, with adaptive algorithm being 'greedier' for higher values). Defaults to zero.}

\item{myopic}{A logical. Does this analysis use 'myopic' response adaptive randomisation (as opposed to 'dynamic' response adaptive randomisation)? Defaults to FALSE.}

\item{u}{A vector of utilities.}
}
\value{

}
\description{
\code{simulateTrial} simulates a decision-theoretic SMART with either fixed randomisation, or 'myopic' or 'dynamic' response adaptive randomisation.
}
\examples{

For a two-stage, two-treatment, degenerate (i.e. participants exit the trial if event) we simulate a participant
with equal randomisation and event rates as follows:

scheme <- data.frame(
stage = c(   1,    1,   2,   2,   2,   2),
int   = c(   1,    1,   1,   1,   1,   1),
a1    = c(   0,    1,   0,   0,   1,   1),
y1    = c(   0,    0,   1,   1,   1,   1),
a2    = c(  NA,   NA,   0,   1,   0,   1),
a1a2  = c(  NA,   NA,   0,   0,   0,   1),
rho   = c( 0.5,  0.5, 0.5, 0.5, 0.5, 0.5),
pi    = c( 0.8,  0.7, 0.5, 0.4, 0.5, 0.4))

utils <- c(0, 0, 0, 0, 1, 1, 1, 1, 1, 1)
simulateTrial(scheme, interimSize = 500, interimNumber = 4, u = utils)

}
